If java.io.tmpdir is mounted with noexec option we change it to /var/tmp during startup. On Unix only of course

+++ gpcore/src/main/java/org/geopublishing/geopublisher/GpUtil.java
@@ -23,6 +23,7 @@ import javax.xml.parsers.FactoryConfigurationError;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.IOFileFilter;
+import org.apache.commons.lang.SystemUtils;
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.apache.log4j.xml.DOMConfigurator;
@@ -247,24 +248,26 @@ public class GpUtil {
 
 	/**
 	 * this method checks for permission on java.io.tmpdir and changes it to a
-	 * given path if files in current tmpDir may not be executed. 
-	 * (Possibly due to noexec mountoption on /tmp)
+	 * given path if files in current tmpDir may not be executed. (Possibly due
+	 * to noexec mountoption on /tmp)
 	 */
 	public static void checkAndResetTmpDir(String path) {
-		String tmpDir = System.getProperty("java.io.tmpdir");
-		try {
-			File file = File.createTempFile("geopublisher", "tmpDirTest",new File(tmpDir));
-			file.setExecutable(true);
-			if(!file.canExecute()){
-				System.setProperty("java.io.tmpdir", path);
-				LOGGER.debug("tmpDir has no execute rights, changing to "+path);
+		if (SystemUtils.IS_OS_UNIX) {
+			String tmpDir = System.getProperty("java.io.tmpdir");
+			try {
+				File file = File.createTempFile("geopublisher", "tmpDirTest",
+						new File(tmpDir));
+				file.setExecutable(true);
+				if (!file.canExecute()) {
+					System.setProperty("java.io.tmpdir", path);
+					LOGGER.debug("tmpDir has no execute rights, changing to "
+							+ path);
+				}
+				FileUtils.deleteQuietly(file);
+			} catch (IOException e) {
+				LOGGER.log(Level.ERROR, e);
 			}
-			FileUtils.deleteQuietly(file);
-		} catch (IOException e) {
-			LOGGER.log(Level.ERROR, e);
-		} finally {
 		}
-
 	}
 
 }
